pyproject.toml
[tool.poetry]
name = "jayrbolton-resume"
version = "0.0.1"
description = ""
authors = ["Jay R Bolton <jayrbolton@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pystache = "^0.5.4"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
